--- dotcom-rendering/src/web/components/Callout/CalloutDescription.tsx
@@ -1,43 +1,0 @@
-import { css } from '@emotion/react';
-import { body, space } from '@guardian/source-foundations';
-import { sanitiseHTML } from '../../../../src/model/sanitise';
-import { decidePalette } from '../../lib/decidePalette';
-
-const descriptionStyles = (format: ArticleFormat) =>
-	css`
-		a {
-			color: ${decidePalette(format).text.richLink};
-			border-bottom: 1px solid ${decidePalette(format).text.richLink};
-			text-decoration: none;
-		}
-		padding-bottom: ${space[4]}px;
-		${body.medium()}
-
-		p {
-			margin-bottom: ${space[3]}px;
-		}
-	`;
-
-interface Props {
-	description: string;
-	format: ArticleFormat;
-}
-
-export const CalloutDescription = ({ description, format }: Props) => (
-	<div css={descriptionStyles(format)}>
-		<div
-			dangerouslySetInnerHTML={{ __html: sanitiseHTML(description) }}
-		></div>
-		<div>
-			Please share your story if you are 18 or over, anonymously if you
-			wish. For more information please see our{' '}
-			<a href="https://www.theguardian.com/help/terms-of-service">
-				terms of service
-			</a>{' '}
-			and{' '}
-			<a href="https://www.theguardian.com/help/privacy-policy">
-				privacy policy
-			</a>
-		</div>
-	</div>
-);
